requirements.txt
boto==2.49.0
boto3==1.17.46
botocore==1.20.78
brotlipy==0.7.0
click==8.0.1
flask==1.1.2
itsdangerous==2.0.1
jinja2==3.0.0
jmespath==0.10.0
markupsafe==2.0.1
pycparser==2.20
pyopenssl==20.0.1
pysocks==1.7.1
python-dateutil==2.8.1
s3transfer==0.3.6
setuptools==52.0.0
six==1.15.0
werkzeug==1.0.1
keras==2.3.1
numpy==1.19.2
pillow==8.1.2 
